menu => Refactoring components and cleanup registration form
