+++ src/assets/main.scss
@@ -29,10 +29,6 @@ body {
     padding: 7px;
   }
 
-  .text-form {
-    color: #434747;
-  }
-
   .link-redirect {
     color: #ffbc1d;
   }
