+++ src/components/VFooter.vue
@@ -15,26 +15,26 @@
 </template>
 
 <style>
-	.footer {
-		position: fixed;
-		left: 0;
-		bottom: 0;
-		width: 100%;
-		text-align: center;
-		background: #161616;
-	}
+.footer {
+  position: fixed;
+  left: 0;
+  bottom: 0;
+  width: 100%;
+  text-align: center;
+  background: #161616;
+}
 
-	.bottom-bar {
-		width: 100%;
-		height: 2px;
-	}
+.bottom-bar {
+  width: 100%;
+  height: 2px;
+}
 
-	.inner .bottom-bar-yellow {
-		width: 100%;
-		height: inherit;
-	}
+.inner .bottom-bar-yellow {
+  width: 100%;
+  height: inherit;
+}
 
-	.bottom-bar-yellow {
-		background: #ffda00;
-	}
+.bottom-bar-yellow {
+  background: #ffda00;
+}
 </style>
