+++ src/components/VMenu.vue
@@ -26,7 +26,7 @@
 			</b-nav>
 		</div>
 
-		<div class="col-md-12 menu" v-else="!isActive">
+		<div class="col-md-12 menu" v-else-if="!isActive">
 			<b-nav vertical>
 				<b-nav-item to="/">
 					<span class="menu-items">
@@ -54,42 +54,40 @@
 </template>
 
 <style>
+.menu-position {
+  background-color: #323231;
+  height: auto;
+  width: auto;
+}
 
-	.menu-position {
-		background-color: #323231;
-		height: auto;
-		width: auto;
-	}
+.button-position {
+  position: relative;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+}
 
-	.button-position {
-		position: relative;
-		bottom: 0;
-		left: 0;
-		width: 100%;
-	}
+.menu-items {
+  color: white;
+}
 
-	.menu-items {
-		color: white;
-	}
+.menu-items:hover {
+  opacity: 0.5;
+}
 
-	.menu-items:hover {
-		opacity: 0.5;
-	}
-
-	.menu {
-		height: 100%;
-		position: fixed;
-		background: #2e2e2d;
-	}
+.menu {
+  height: 100%;
+  position: fixed;
+  background: #2e2e2d;
+}
 </style>
 
 <script>
-  export default {
-    data: function () {
-      return {
-        isActive: true
-      }
-    }
+export default {
+  data: function() {
+    return {
+      isActive: true
+    };
   }
+};
 </script>
-
